Fail with informative error if explicit --tmp-dir isn't writable
Previously, if the passed --tmp-dir did not exist or was not writable,
cachetool would fallback to /dev/shm. As the user is explicitly passing
a directory, it is safe to assume that they want to use that directory
and only that directly. Falling back to some other directory is
undesired and counter intuitive. In my own case, it made debugging more
difficult as I had difficulty correlating directory permissions with
expected results.

Now assumes the passed --tmp-dir and if it is not writable, will fail
appropriately with a permission error or "No such file or directory"
error.