Support adding host header to web adapter
When using the Symfony HTTP client as web adapter it is now possible
to set the "Host:" header to the request made by cachetool. This
allows using URLs with specific IP addresses as host part, specifying
the actual hostname as a HTTP header.

This makes it possible to use cachetool on load-balanced web servers.

See https://github.com/gordalina/cachetool/issues/162