Parse --web option correctly
`getParameterOption` returns the unparsed / unvalidated result of the option,
which might result in incorrect results when reordering options:

```
php cachetool.phar opcache:status --web --web-path=/var/www
```

returns --web-path=/var/www as the value for the --web option when using
`getParameterOption`, getting the parsed result is done by using `getOption`
which returns the correctly parsed value, in this case `NULL`, which should
result in the default being set which is FileGetContents.

This avoids the warning that the webclient is unkown when passing --web without
specifying the client.